chore: 코드 청소

+++ src/routes/habits/habits.routes.js
@@ -19,6 +19,12 @@ habitsRouter.post(
       const { id } = req.params;
       const { checkDate } = req.body;
 
+      const habit = await habitsRepository.findHabitById(id);
+
+      if (!habit) {
+        throw new NotFoundException(ERROR_MESSAGE.HABIT_NOT_FOUND);
+      }
+
       const newHabitCheckDate = await habitsRepository.toggleHabitCheckDate(
         id,
         checkDate,
@@ -43,6 +49,12 @@ habitsRouter.delete(
       const { id } = req.params;
       const { checkDate } = req.body;
 
+      const habit = await habitsRepository.findHabitById(id);
+
+      if (!habit) {
+        throw new NotFoundException(ERROR_MESSAGE.HABIT_NOT_FOUND);
+      }
+
       await habitsRepository.deleteHabitCheckDate(id, checkDate);
 
       res.status(HTTP_STATUS.NO_CONTENT);
@@ -51,53 +63,3 @@ habitsRouter.delete(
     }
   },
 );
-
-// 특정 Study의 습관을 삭제/신규/수정을 동기화 처리하기 위한 로직
-// PUT /studies/:studyId
-// habitsRouter.put(
-//   '/:id', //studyId
-//   validateObject(habitsSchema.params, 'params'),
-//   validateObject(habitsSchema.body, 'body'),
-//   async (req, res, next) => {
-//     try {
-//       const { id: studyId } = req.params;
-//       const habits = req.body;
-
-//       await prisma.$transaction(async (tx) => {
-//         //기존 Habit중 현재 isDelted:False 전체습관 조회
-//         const existingHabits = await habitsRepository.findActiveByStudyId(
-//           tx,
-//           studyId,
-//         );
-
-//         // 추가: 요청받은 데이터 중 유효한(숫자) Habit Id 목록 추출
-//         const incomingIds = habits.map((h) => h.id).filter(Boolean);
-
-//         // Delted 처리할 대상을 선별 - 새로 넘어오지 않은 Habits 추출(isDeleted : true처리 목적)
-//         const habitsToDelete = existingHabits.filter(
-//           (existingHabit) =>
-//             !incomingIds.includes(existingHabit.id),
-//         );
-
-//         // 생성 처리 대상 구분 : id가 없는(null) 표시되어온 대상 : 신규입력 대상
-//         const habitsToCreate = habits.filter((habit) => !habit.id);
-
-//         //습관수정 대상 - FE로부터 아무 표시가 없는 대상 : name 수정 대상
-//         const habitsToUpdate = habits.filter(
-//           (habit) => habit.id,
-//         );
-
-//         // 삭제/신규/수정 일괄 처리
-//         await Promise.all([
-//           habitsRepository.deleteHabits(tx, habitsToDelete),
-//           habitsRepository.createHabits(tx, studyId, habitsToCreate),
-//           habitsRepository.updateHabits(tx, habitsToUpdate),
-//         ]);
-//       });
-
-//       res.sendStatus(HTTP_STATUS.NO_CONTENT);
-//     } catch (error) {
-//       next(error);
-//     }
-//   },
-// );
